build: change JavaScript target to ES2022
Thus, the target is the most modern version of JavaScript that is fully supported by Node.js v18+.
(No older Node.js version should be used nowadays.)

Now, this library can be used in a modern environment as is.
To be used in older environments, a transpiler such as Babel must be used.